requirements.txt
PyQt6>=6.5
PyYAML>=6.0
